fix: centralize objecxt constructor in interface
